+++ components/modals/edit-mentor-modal/tabs/index.ts
@@ -19,3 +19,4 @@ export * from './audit-log-tab';
 export * from './voice-tab';
 export * from './screenshare-tab';
 export * from './lti-tab';
+export * from './analytics-tab';